Move proof functions into ProofTrie subclass
